+++ Exodus-Test-Server/src/utopia/exodus/test/rest/servlet/ApiServlet.scala
@@ -16,6 +16,7 @@ import utopia.flow.generic.DataType
 import utopia.flow.parse.JsonParser
 import utopia.flow.util.StringExtensions.ExtendedString
 import utopia.flow.util.FileExtensions._
+import utopia.flow.util.logging.{Logger, SysErrLogger}
 import utopia.nexus.http.{Path, ServerSettings}
 import utopia.nexus.rest.RequestHandler
 import utopia.vault.database.{Connection, ConnectionPool}
@@ -44,6 +45,7 @@ class ApiServlet extends HttpServlet
 	
 	DataType.setup()
 	
+	private implicit val logger: Logger = SysErrLogger
 	private implicit val exc: ExecutionContext = new ThreadPool("Exodus-Test-Server").executionContext
 	private implicit val connectionPool: ConnectionPool = new ConnectionPool()
 	
@@ -51,12 +53,11 @@ class ApiServlet extends HttpServlet
 	val dbSettings = dbSettingsRead.getOrElse(Model.empty)
 	
 	ExodusContext.setup(exc, connectionPool,
-		dbSettings("db_name", "db").stringOr("exodus_db")) { (error, message) =>
-		println(message)
-		error.printStackTrace()
-	} { Set(ReadGeneralData, ReadPersonalData, PersonalActions, ReadOrganizationData, OrganizationActions,
+		dbSettings("db_name", "db").stringOr("exodus_db")) {
+		Set(ReadGeneralData, ReadPersonalData, PersonalActions, ReadOrganizationData, OrganizationActions,
 		CreateOrganization, RequestPasswordReset, ChangeKnownPassword, TerminateOtherSessions, RevokeOtherTokens,
-		JoinOrganization) }
+		JoinOrganization)
+	}
 	Connection.modifySettings { _.copy(driver = Some("org.mariadb.jdbc.Driver"), charsetName = "utf8",
 		charsetCollationName = "utf8_general_ci") }
 	dbSettingsRead match